refactor: simplify imports by consolidating component paths
