+++ src/app/profile/[username]/[repoName]/500.tsx
@@ -1,4 +1,4 @@
-import { ErrorFrame } from "@/components/Frames/ErrorFrame";
+import { ErrorFrame } from "@/components";
 
 const custom500 = () => {
   return (
