+++ src/components/Frames/index.ts
@@ -1,1 +1,3 @@
 export * from "./ProfileFrame"
+export * from "./ErrorFrame"
+export * from "./RepositoryFrame"
